refactor: move tools and ids in their own components
