+++ apps/web-tools/package.json
@@ -16,6 +16,7 @@
     "typecheck": "tsc --noEmit --customConditions node"
   },
   "dependencies": {
+    "@devtools/feature-flags": "workspace:*",
     "@devtools/shell": "workspace:*",
     "@reduxjs/toolkit": "catalog:",
     "@shared/feature-flags": "workspace:*",
