+++ devtools/feature-flags/src/index.ts
@@ -1,5 +1,6 @@
 import "./register";
 
+export { FEATURE_FLAGS_ID } from "./register";
 export type { FeatureFlagsToolProps, FlagDisplayState, FlagFilter } from "./types";
 export { ALL_FLAG_IDS } from "./constants";
 export { useFeatureFlagsState } from "./hooks";
