--- devtools/shell/src/toolIds.ts
@@ -1,1 +1,0 @@
-export const FEATURE_FLAGS_ID = "feature-flags" as const;
